backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
black==25.1.0
boto3==1.40.30
botocore==1.40.30
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
filelock==3.13.1
flake8==7.3.0
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
jq==1.10.0
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.3
motor==3.3.1
mpmath==1.3.0
mypy==1.18.1
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.3
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.4.0
pluggy==1.6.0
protobuf==6.32.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.8.0+cpu
torchaudio==2.8.0
torchvision==0.23.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0
Werkzeug==3.1.3
wrapt==1.17.3
